fix: column id_pgto

+++ database/migrations/2025_02_08_220354_create_doacoes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('ip');
             $table->string('email')->default('email@email.com');
             $table->double('valor');
-            $table->integer('id_pgto');
+            $table->bigInteger('id_pgto');
             $table->string('idempotence');
             $table->string('status');
             $table->timestamps();
